chore(web): delete dead components and the use-debounce re-export shim
Removes unreferenced code found in the codebase completeness review,
each deletion verified against direct, aliased, and relative imports:

- Dead components: HouseIcon, StarIcon, currency-converter, nav-user,
  leatest-change, logo, the unused animate-ui tabs pair, and the
  shadcn-space number-ticker vendor file.
- The api-keys analytics mock dataset (fabricated sparkline data); the
  file keeps its real exports and is renamed to analytics-data.ts. The
  unreachable isLoading branch under useSuspenseQuery is removed.
- apps/web/src/hooks/use-debounce.ts was a one-line re-export of
  @trenova/shared/hooks/use-debounce; importers and test mocks now use
  the shared module directly.

The audit had also flagged header, keyboard-shortcuts-dialog,
recommended-badge, system-information, and form-copy-button as dead;
they are referenced through relative imports and are kept.
