--- client/apps/web/src/components/HouseIcon.tsx
@@ -1,103 +1,0 @@
-import { cn } from "@trenova/shared/lib/utils";
-import type { HTMLMotionProps, Variants } from "motion/react";
-import { m, useAnimation, useReducedMotion } from "motion/react";
-import { forwardRef, useCallback, useImperativeHandle, useRef } from "react";
-
-export interface HouseHandle {
-  startAnimation: () => void;
-  stopAnimation: () => void;
-}
-
-interface HouseProps extends HTMLMotionProps<"div"> {
-  size?: number;
-  duration?: number;
-  isAnimated?: boolean;
-}
-
-const HouseIcon = forwardRef<HouseHandle, HouseProps>(
-  (
-    { onMouseEnter, onMouseLeave, className, size = 24, duration = 1, isAnimated = true, ...props },
-    ref,
-  ) => {
-    const controls = useAnimation();
-    const reduced = useReducedMotion();
-    const isControlled = useRef(false);
-
-    useImperativeHandle(ref, () => {
-      isControlled.current = true;
-      return {
-        startAnimation: () => (reduced ? controls.start("normal") : controls.start("animate")),
-        stopAnimation: () => controls.start("normal"),
-      };
-    });
-
-    const handleEnter = useCallback(
-      (e?: React.MouseEvent<HTMLDivElement>) => {
-        if (!isAnimated || reduced) return;
-        if (!isControlled.current) void controls.start("animate");
-        else onMouseEnter?.(e as any);
-      },
-      [controls, reduced, isAnimated, onMouseEnter],
-    );
-
-    const handleLeave = useCallback(
-      (e?: React.MouseEvent<HTMLDivElement>) => {
-        if (!isControlled.current) void controls.start("normal");
-        else onMouseLeave?.(e as any);
-      },
-      [controls, onMouseLeave],
-    );
-
-    const baseVariants: Variants = {
-      normal: { opacity: 1 },
-      animate: {
-        opacity: 0.65,
-        transition: {
-          duration: 0.2 * duration,
-          ease: "easeOut",
-        },
-      },
-    };
-
-    const doorVariants: Variants = {
-      normal: { opacity: 1 },
-      animate: {
-        opacity: [1, 0.4, 1],
-        transition: {
-          duration: 0.35 * duration,
-          ease: "easeInOut",
-        },
-      },
-    };
-
-    return (
-      <m.div
-        className={cn("inline-flex items-center justify-center", className)}
-        onMouseEnter={handleEnter}
-        onMouseLeave={handleLeave}
-        {...props}
-      >
-        <m.svg
-          xmlns="http://www.w3.org/2000/svg"
-          width={size}
-          height={size}
-          viewBox="0 0 24 24"
-          fill="none"
-          stroke="currentColor"
-          strokeWidth="2"
-          strokeLinecap="round"
-          strokeLinejoin="round"
-          animate={controls}
-          initial="normal"
-        >
-          <path d="M3 10a2 2 0 0 1 .709-1.528l7-5.999a2 2 0 0 1 2.582 0l7 5.999A2 2 0 0 1 21 10" />
-          <m.path d="M21 10v9a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2v-9" variants={baseVariants} />
-          <m.path d="M15 21v-8a1 1 0 0 0-1-1h-4a1 1 0 0 0-1 1v8" variants={doorVariants} />
-        </m.svg>
-      </m.div>
-    );
-  },
-);
-
-HouseIcon.displayName = "HouseIcon";
-export { HouseIcon };
