--- client/apps/web/src/components/StarIcon.tsx
@@ -1,111 +1,0 @@
-import { cn } from "@trenova/shared/lib/utils";
-import type { HTMLMotionProps, Variants } from "motion/react";
-import { m, useAnimation, useReducedMotion } from "motion/react";
-import { forwardRef, useCallback, useImperativeHandle, useRef } from "react";
-
-export interface StarIconHandle {
-  startAnimation: () => void;
-  stopAnimation: () => void;
-}
-
-interface StarIconProps extends HTMLMotionProps<"div"> {
-  size?: number;
-  duration?: number;
-  isAnimated?: boolean;
-  strokeWidth?: number;
-  fill?: string;
-}
-
-const StarIcon = forwardRef<StarIconHandle, StarIconProps>(
-  (
-    {
-      onMouseEnter,
-      onMouseLeave,
-      className,
-      size = 24,
-      duration = 1,
-      isAnimated = true,
-      strokeWidth = 2,
-      fill = "currentColor",
-      ...props
-    },
-    ref,
-  ) => {
-    const controls = useAnimation();
-    const reduced = useReducedMotion();
-    const isControlled = useRef(false);
-
-    useImperativeHandle(ref, () => {
-      isControlled.current = true;
-      return {
-        startAnimation: () => (reduced ? controls.start("normal") : controls.start("animate")),
-        stopAnimation: () => controls.start("normal"),
-      };
-    });
-
-    const handleEnter = useCallback(
-      (e?: React.MouseEvent<HTMLDivElement>) => {
-        if (!isAnimated || reduced) return;
-        if (!isControlled.current) void controls.start("animate");
-        else onMouseEnter?.(e as any);
-      },
-      [controls, reduced, isAnimated, onMouseEnter],
-    );
-
-    const handleLeave = useCallback(
-      (e?: React.MouseEvent<HTMLDivElement>) => {
-        if (!isControlled.current) void controls.start("normal");
-        else onMouseLeave?.(e as any);
-      },
-      [controls, onMouseLeave],
-    );
-
-    const starVariants: Variants = {
-      idle: {
-        scale: [1, 1.02, 0.98, 1],
-        transition: {
-          duration: 2.5 * duration,
-          repeat: Infinity,
-          ease: "easeInOut",
-        },
-      },
-      animate: {
-        scale: [1, 1.2, 0.95, 1.05, 1],
-        rotate: [0, -10, 10, 0],
-        transition: { duration: 1.2 * duration, ease: "easeInOut" },
-      },
-    };
-
-    return (
-      <m.div
-        className={cn("inline-flex", className)}
-        onMouseEnter={handleEnter}
-        onMouseLeave={handleLeave}
-        {...props}
-      >
-        <m.svg
-          xmlns="http://www.w3.org/2000/svg"
-          width={size}
-          height={size}
-          viewBox="0 0 24 24"
-          fill={fill}
-          stroke="currentColor"
-          strokeWidth={strokeWidth}
-          strokeLinecap="round"
-          strokeLinejoin="round"
-          animate={controls}
-          initial="idle"
-          variants={starVariants}
-        >
-          <m.path
-            d="M11.525 2.295a.53.53 0 0 1 .95 0l2.31 4.679a2.123 2.123 0 0 0 1.595 1.16l5.166.756a.53.53 0 0 1 .294.904l-3.736 3.638a2.123 2.123 0 0 0-.611 1.878l.882 5.14a.53.53 0 0 1-.771.56l-4.618-2.428a2.122 2.122 0 0 0-1.973 0L6.396 21.01a.53.53 0 0 1-.77-.56l.881-5.139a2.122 2.122 0 0 0-.611-1.879L2.16 9.795a.53.53 0 0 1 .294-.906l5.165-.755a2.122 2.122 0 0 0 1.597-1.16z"
-            fill={fill}
-          />
-        </m.svg>
-      </m.div>
-    );
-  },
-);
-
-StarIcon.displayName = "StarIcon";
-export { StarIcon };
