--- client/apps/web/src/components/animate-ui/components/base/tabs.tsx
@@ -1,78 +1,0 @@
-import {
-  TabsHighlightItem as TabsHighlightItemPrimitive,
-  TabsHighlight as TabsHighlightPrimitive,
-  TabsList as TabsListPrimitive,
-  TabsPanel as TabsPanelPrimitive,
-  TabsPanels as TabsPanelsPrimitive,
-  Tabs as TabsPrimitive,
-  TabsTab as TabsTabPrimitive,
-  type TabsListProps as TabsListPrimitiveProps,
-  type TabsPanelProps as TabsPanelPrimitiveProps,
-  type TabsPanelsProps as TabsPanelsPrimitiveProps,
-  type TabsProps as TabsPrimitiveProps,
-  type TabsTabProps as TabsTabPrimitiveProps,
-} from "@/components/animate-ui/primitives/base/tabs";
-import { cn } from "@trenova/shared/lib/utils";
-
-type TabsProps = TabsPrimitiveProps;
-
-function Tabs({ className, ...props }: TabsProps) {
-  return <TabsPrimitive className={cn("flex flex-col gap-2", className)} {...props} />;
-}
-
-type TabsListProps = TabsListPrimitiveProps;
-
-function TabsList({ className, ...props }: TabsListProps) {
-  return (
-    <TabsHighlightPrimitive className="bg-background dark:border-input dark:bg-input/30 absolute inset-0 z-0 rounded-md border border-transparent shadow-sm">
-      <TabsListPrimitive
-        className={cn(
-          "bg-muted text-muted-foreground inline-flex h-9 w-fit items-center justify-center rounded-lg p-[3px]",
-          className,
-        )}
-        {...props}
-      />
-    </TabsHighlightPrimitive>
-  );
-}
-
-type TabsTabProps = TabsTabPrimitiveProps;
-
-function TabsTab({ className, ...props }: TabsTabProps) {
-  return (
-    <TabsHighlightItemPrimitive value={props.value} className="flex-1">
-      <TabsTabPrimitive
-        className={cn(
-          "text-muted-foreground focus-visible:border-ring focus-visible:ring-ring/50 focus-visible:outline-ring data-[selected]:text-foreground inline-flex h-[calc(100%-1px)] w-full flex-1 items-center justify-center gap-1.5 rounded-md px-2 py-1 text-sm font-medium whitespace-nowrap transition-colors duration-500 ease-in-out focus-visible:ring-[3px] focus-visible:outline-1 disabled:pointer-events-none disabled:opacity-50 [&_svg]:pointer-events-none [&_svg]:shrink-0 [&_svg:not([class*='size-'])]:size-4",
-          className,
-        )}
-        {...props}
-      />
-    </TabsHighlightItemPrimitive>
-  );
-}
-
-type TabsPanelsProps = TabsPanelsPrimitiveProps;
-
-function TabsPanels(props: TabsPanelsProps) {
-  return <TabsPanelsPrimitive {...props} />;
-}
-
-type TabsPanelProps = TabsPanelPrimitiveProps;
-
-function TabsPanel({ className, ...props }: TabsPanelProps) {
-  return <TabsPanelPrimitive className={cn("flex-1 outline-none", className)} {...props} />;
-}
-
-export {
-  Tabs,
-  TabsList,
-  TabsPanel,
-  TabsPanels,
-  TabsTab,
-  type TabsListProps,
-  type TabsPanelProps,
-  type TabsPanelsProps,
-  type TabsProps,
-  type TabsTabProps,
-};
