--- client/apps/web/src/components/shadcn-space/number-ticker/number-ticker-02.tsx
@@ -1,30 +1,0 @@
-import NumberFlow, { type Value } from "@number-flow/react";
-
-type NumberTickerProps = {
-  value: Value;
-  currency?: string;
-  decimals?: number;
-  className?: string;
-};
-
-export function NumberTicker({
-  value,
-  currency = "USD",
-  decimals = 2,
-  className,
-}: NumberTickerProps) {
-  return (
-    <div className="inline-flex items-center gap-3">
-      <NumberFlow
-        value={value}
-        format={{
-          style: "currency",
-          currency: currency,
-          minimumFractionDigits: decimals,
-          maximumFractionDigits: decimals,
-        }}
-        className={className}
-      />
-    </div>
-  );
-}
