+++ client/apps/web/src/hooks/__tests__/use-shipment-totals-preview.test.tsx
@@ -16,7 +16,7 @@ vi.mock("@/services/api", () => ({
 }));
 
 // The debounce the hook puts in front of every request.
-vi.mock("@/hooks/use-debounce", () => ({
+vi.mock("@trenova/shared/hooks/use-debounce", () => ({
   useDebounce: <T,>(value: T) => value,
 }));
