+++ client/apps/web/src/routes/admin/api-keys/_components/analytics/api-key-analytics.tsx
@@ -1,21 +1,14 @@
-import { MetricSkeleton } from "@/components/metric-skeleton";
 import { analytics } from "@/lib/queries/analytics";
 import { useSuspenseQuery } from "@tanstack/react-query";
 import { ActiveKeysCard } from "./cards/active-keys-card";
 import { RequestsCard } from "./cards/requests-card";
 import { RevokedKeysCard } from "./cards/revoked-keys-card";
 import { TotalKeysCard } from "./cards/total-keys-card";
-import { defaultApiKeyAnalyticsData, type ApiKeyAnalyticsData } from "./mock-data";
+import { defaultApiKeyAnalyticsData, type ApiKeyAnalyticsData } from "./analytics-data";
 
 export default function ApiKeyAnalytics() {
-  const { data: raw, isLoading: dataLoading } = useSuspenseQuery(
-    analytics.get("api-key-management"),
-  );
+  const { data: raw } = useSuspenseQuery(analytics.get("api-key-management"));
   const data = (raw as ApiKeyAnalyticsData) ?? defaultApiKeyAnalyticsData;
-
-  if (dataLoading) {
-    return <MetricSkeleton length={4} />;
-  }
 
   return (
     <div className="grid grid-cols-2 gap-2.5 lg:grid-cols-4 pb-3">
