+++ client/apps/web/src/routes/service-type/_components/__tests__/service-type-table.test.tsx
@@ -68,7 +68,7 @@ vi.mock("@/lib/queries", () => ({
   },
 }));
 
-vi.mock("@/hooks/use-debounce", () => ({
+vi.mock("@trenova/shared/hooks/use-debounce", () => ({
   useDebounce: <T,>(value: T): T => value,
 }));
